drf_backend/requirements.txt
Django
djangorestframework
markdown
python-dotenv
django-cors-headers
psycopg2-binary
Pillow
drf-yasg
django-filter
whitenoise
gunicorn
django-environ
djangorestframework-simplejwt
django-storages
six
PyJWT
pytz
dj-database-url
python-decouple
dj-rest-auth
django-allauth
social-auth-app-django